qd_sample

qd_sample

qd_sample(categories, size)

Arguments

categories
a vector of valid QuickDraw category

size
a vector of number of rows to take from each category

Value

A tibble object that binds the sample data from all catgories.

Details

qd_sample() downloads QuickDraw data from multiple categories and take a sample from each category

Examples

qd_sample(c("octopus", "star"), 5)

#> QuickDraw file: octopus.ndjson has been downloaded already!

#> QuickDraw file: star.ndjson has been downloaded already!

#> # A tibble: 10 x 6
#>    word    countrycode timestamp               recognized key_id       drawing  
#>    <chr>   <chr>       <chr>                   <lgl>      <chr>        <list>   
#>  1 octopus SG          2017-01-23 14:20:39.82… TRUE       60902023658… <list [6…
#>  2 octopus US          2017-03-02 03:07:06.89… TRUE       51149825239… <list [2…
#>  3 octopus US          2017-03-02 00:37:31.36… TRUE       55504148983… <list [1…
#>  4 octopus US          2017-03-03 04:52:17.48… TRUE       48966381978… <list [1…
#>  5 octopus GB          2017-03-26 20:38:44.24… TRUE       59473296537… <list [1…
#>  6 star    BR          2017-01-28 21:08:18.88… TRUE       53435616607… <list [1…
#>  7 star    US          2017-03-12 21:00:35.83… TRUE       58481941413… <list [3…
#>  8 star    JP          2017-03-13 05:43:37.03… TRUE       66015024140… <list [4…
#>  9 star    US          2017-03-09 18:00:47.92… TRUE       64917631810… <list [1…
#> 10 star    US          2017-03-23 20:42:44.45… TRUE       60174532492… <list [2…